Move check list to MC controller
