+++ _24hplusdotnetcore/Controllers/CheckInfoController.cs
@@ -103,29 +103,5 @@ namespace _24hplusdotnetcore.Controllers
             }
         }
 
-        [HttpGet]
-        [Route("api/check-list")]
-        public async Task<ActionResult<ResponseContext>> CheckListAsync([FromQuery] string customerId)
-        {
-            try
-            {
-                CustomerCheckListResponseModel result = await _checkInforServices.CheckListAsync(customerId);
-                return Ok(new ResponseContext
-                {
-                    code = (int)Common.ResponseCode.SUCCESS,
-                    message = Common.Message.SUCCESS,
-                    data = result
-                });
-            }
-            catch (Exception ex)
-            {
-                _logger.LogError(ex, ex.Message);
-                return StatusCode(StatusCodes.Status500InternalServerError, new ResponseMessage
-                {
-                    status = "ERROR",
-                    message = ex.Message
-                });
-            }
-        }
     }
 }
