+++ _24hplusdotnetcore/Services/CheckInfoServices.cs
@@ -13,19 +13,13 @@ namespace _24hplusdotnetcore.Services
     {
         private readonly ILogger<CheckInfoServices> _logger;
         private readonly MCService _mcServices;
-        private readonly IRestMCService _restMCService;
-        private readonly CustomerServices _customerServices;
 
         public CheckInfoServices(
             ILogger<CheckInfoServices> logger,
-            MCService mcServices,
-            IRestMCService restMCService,
-            CustomerServices customerServices)
+            MCService mcServices)
         {
             _logger = logger;
             _mcServices = mcServices;
-            _restMCService = restMCService;
-            _customerServices = customerServices;
         }
         public dynamic CheckInfoByType(string greentype, string citizenID, string customerName)
         {
@@ -162,21 +156,6 @@ namespace _24hplusdotnetcore.Services
                 _logger.LogError(ex, ex.Message);
             }
             return token;
-        }
-
-        public async Task<CustomerCheckListResponseModel> CheckListAsync(string customerId)
-        {
-            try
-            {
-                CustomerCheckListRequestModel customerCheckList = await _customerServices.GetCustomerCheckListAsync(customerId);
-                CustomerCheckListResponseModel result = await _restMCService.CheckListAsync(customerCheckList);
-                return result;
-            }
-            catch (Exception ex)
-            {
-                _logger.LogError(ex, ex.Message);
-                throw;
-            }
         }
     }
 }
